Move Invocation def into separate file
