+++ cmd/ipfs/invocation.go
@@ -1,0 +1,132 @@
+package main
+
+import (
+	"errors"
+	"io"
+	"os"
+
+	context "github.com/jbenet/go-ipfs/Godeps/_workspace/src/code.google.com/p/go.net/context"
+	cmds "github.com/jbenet/go-ipfs/commands"
+	cmdsCli "github.com/jbenet/go-ipfs/commands/cli"
+	core "github.com/jbenet/go-ipfs/core"
+	fsrepo "github.com/jbenet/go-ipfs/repo/fsrepo"
+	u "github.com/jbenet/go-ipfs/util"
+)
+
+type cmdInvocation struct {
+	path []string
+	cmd  *cmds.Command
+	req  cmds.Request
+	node *core.IpfsNode
+}
+
+func (i *cmdInvocation) Run(ctx context.Context) (output io.Reader, err error) {
+	// setup our global interrupt handler.
+	i.setupInterruptHandler()
+
+	// check if user wants to debug. option OR env var.
+	debug, _, err := i.req.Option("debug").Bool()
+	if err != nil {
+		return nil, err
+	}
+	if debug || u.GetenvBool("DEBUG") || os.Getenv("IPFS_LOGGING") == "debug" {
+		u.Debug = true
+		u.SetDebugLogging()
+	}
+
+	res, err := callCommand(ctx, i.req, Root, i.cmd)
+	if err != nil {
+		return nil, err
+	}
+
+	if err := res.Error(); err != nil {
+		return nil, err
+	}
+
+	return res.Reader()
+}
+
+func (i *cmdInvocation) constructNodeFunc(ctx context.Context) func() (*core.IpfsNode, error) {
+	return func() (*core.IpfsNode, error) {
+		if i.req == nil {
+			return nil, errors.New("constructing node without a request")
+		}
+
+		cmdctx := i.req.Context()
+		if cmdctx == nil {
+			return nil, errors.New("constructing node without a request context")
+		}
+
+		r := fsrepo.At(i.req.Context().ConfigRoot)
+		if err := r.Open(); err != nil { // repo is owned by the node
+			return nil, err
+		}
+
+		// ok everything is good. set it on the invocation (for ownership)
+		// and return it.
+		n, err := core.NewIPFSNode(ctx, core.Standard(r, cmdctx.Online))
+		if err != nil {
+			return nil, err
+		}
+		i.node = n
+		return i.node, nil
+	}
+}
+
+func (i *cmdInvocation) close() {
+	// let's not forget teardown. If a node was initialized, we must close it.
+	// Note that this means the underlying req.Context().Node variable is exposed.
+	// this is gross, and should be changed when we extract out the exec Context.
+	if i.node != nil {
+		log.Info("Shutting down node...")
+		i.node.Close()
+	}
+}
+
+func (i *cmdInvocation) Parse(ctx context.Context, args []string) error {
+	var err error
+
+	i.req, i.cmd, i.path, err = cmdsCli.Parse(args, os.Stdin, Root)
+	if err != nil {
+		return err
+	}
+	i.req.Context().Context = ctx
+
+	repoPath, err := getRepoPath(i.req)
+	if err != nil {
+		return err
+	}
+	log.Debugf("config path is %s", repoPath)
+
+	// this sets up the function that will initialize the config lazily.
+	cmdctx := i.req.Context()
+	cmdctx.ConfigRoot = repoPath
+	cmdctx.LoadConfig = loadConfig
+	// this sets up the function that will initialize the node
+	// this is so that we can construct the node lazily.
+	cmdctx.ConstructNode = i.constructNodeFunc(ctx)
+
+	// if no encoding was specified by user, default to plaintext encoding
+	// (if command doesn't support plaintext, use JSON instead)
+	if !i.req.Option("encoding").Found() {
+		if i.req.Command().Marshalers != nil && i.req.Command().Marshalers[cmds.Text] != nil {
+			i.req.SetOption("encoding", cmds.Text)
+		} else {
+			i.req.SetOption("encoding", cmds.JSON)
+		}
+	}
+
+	return nil
+}
+
+func (i *cmdInvocation) requestedHelp() (short bool, long bool, err error) {
+	longHelp, _, err := i.req.Option("help").Bool()
+	if err != nil {
+		return false, false, err
+	}
+	shortHelp, _, err := i.req.Option("h").Bool()
+	if err != nil {
+		return false, false, err
+	}
+	return longHelp, shortHelp, nil
+}
